Blob unit tests updates and refactor
