--- test/WorkerExtensionTests/Blob/BlobStorageConverterTests.cs
@@ -1,1004 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the MIT License. See License.txt in the project root for license information.
-
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading;
-using System.Threading.Tasks;
-using Azure;
-using Azure.Storage.Blobs;
-using Azure.Storage.Blobs.Models;
-using Microsoft.Azure.Functions.Worker;
-using Microsoft.Azure.Functions.Worker.Converters;
-using Microsoft.Azure.Functions.Worker.Tests.Converters;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Options;
-using Moq;
-using Xunit;
-
-// Scenarios for BlobBaseClient, BlockBlobClient, PageBlobClient, and AppendBlobClient
-// are tested via E2E tests as Moq does not support mocking extension methods directly
-namespace Microsoft.Azure.Functions.WorkerExtension.Tests
-{
-    public class BlobStorageConverterTests
-    {
-        private readonly BlobStorageConverter _blobStorageConverter;
-        private readonly Mock<BlobServiceClient> _mockBlobServiceClient;
-
-        public BlobStorageConverterTests()
-        {
-            var host = new HostBuilder().ConfigureFunctionsWorkerDefaults((WorkerOptions options) => { }).Build();
-
-            var workerOptions = host.Services.GetService<IOptions<WorkerOptions>>();
-            var logger = host.Services.GetService<ILogger<BlobStorageConverter>>();
-
-            _mockBlobServiceClient = new Mock<BlobServiceClient>();
-
-            var mockBlobOptions = new Mock<BlobStorageBindingOptions>();
-            mockBlobOptions.Object.Client = _mockBlobServiceClient.Object;
-
-            var mockBlobOptionsSnapshot = new Mock<IOptionsSnapshot<BlobStorageBindingOptions>>();
-            mockBlobOptionsSnapshot
-                .Setup(m => m.Get(It.IsAny<string>()))
-                .Returns(mockBlobOptions.Object);
-
-            _blobStorageConverter = new BlobStorageConverter(workerOptions, mockBlobOptionsSnapshot.Object, logger);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_BlobClient_FilePath_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient.Setup(m => m.Name).Returns("MyBlob");
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var clientResult = (BlobClient)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.Equal("MyBlob", clientResult.Name);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_BlobClientCollection_List_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(IEnumerable<BlobClient>), grpcModelBindingData);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient.Setup(m => m.Name).Returns("MyBlob");
-
-            var mockResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var clientResult = (IEnumerable<BlobClient>)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsAssignableFrom<IEnumerable<BlobClient>>(clientResult);
-            Assert.Equal("MyBlob", clientResult.First().Name);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_BlobClientCollection_Array_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient[]), grpcModelBindingData);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient.Setup(m => m.Name).Returns("MyBlob");
-
-            var mockResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var clientResult = (BlobClient[])conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsType<BlobClient[]>(clientResult);
-            Assert.Equal("MyBlob", clientResult.First().Name);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_BlobClientCollection_WithSubdirectoryPath_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData("test"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient[]), grpcModelBindingData);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient.Setup(m => m.Name).Returns("MyBlob");
-
-            var mockResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var clientResult = (BlobClient[])conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsType<BlobClient[]>(clientResult);
-            Assert.Equal("MyBlob", clientResult.First().Name);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_BlobContainerClient_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobContainerClient), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.Name).Returns("MyContainer");
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var clientResult = (BlobContainerClient)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.Equal("MyContainer", clientResult.Name);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_String_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(string), grpcModelBindingData);
-
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadResult(new BinaryData("MyBlobString"));
-            var mockResponse = new Mock<Response<BlobDownloadResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient.Setup(m => m.DownloadContentAsync()).ReturnsAsync(mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var stringResult = (string)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.Equal("MyBlobString", stringResult);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_StringCollection_List_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(IEnumerable<string>), grpcModelBindingData);
-
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadResult(new BinaryData("MyBlobString"));
-            var mockResponse = new Mock<Response<BlobDownloadResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient.Setup(m => m.DownloadContentAsync()).ReturnsAsync(mockResponse.Object);
-
-            var mockBlobItemResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockBlobItemResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var stringResult = (IEnumerable<string>)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsAssignableFrom<IEnumerable<string>>(stringResult);
-            Assert.Equal("MyBlobString", stringResult.First().ToString());
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_StringCollection_Array_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(string[]), grpcModelBindingData);
-
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadResult(new BinaryData("MyBlobString"));
-            var mockResponse = new Mock<Response<BlobDownloadResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient.Setup(m => m.DownloadContentAsync()).ReturnsAsync(mockResponse.Object);
-
-            var mockBlobItemResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockBlobItemResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var stringResult = (string[])conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsType<string[]>(stringResult);
-            Assert.Equal("MyBlobString", stringResult.First().ToString());
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_Stream_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(Stream), grpcModelBindingData);
-
-            var expectedStream = new MemoryStream(Encoding.UTF8.GetBytes("MyBlobStream"));
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadStreamingResult(expectedStream);
-            var mockResponse = new Mock<Response<BlobDownloadStreamingResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadStreamingAsync(It.IsAny<HttpRange>(), It.IsAny<BlobRequestConditions>(), It.IsAny<bool>(), null, default))
-                .ReturnsAsync(mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var streamResult = (Stream)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.Equal(expectedStream, streamResult);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_StreamCollection_List_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(IEnumerable<Stream>), grpcModelBindingData);
-
-            var expectedStream = new MemoryStream(Encoding.UTF8.GetBytes("MyBlobStream"));
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadStreamingResult(expectedStream);
-
-            var mockResponse = new Mock<Response<BlobDownloadStreamingResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadStreamingAsync(It.IsAny<HttpRange>(), It.IsAny<BlobRequestConditions>(), It.IsAny<bool>(), null, default))
-                .ReturnsAsync(mockResponse.Object);
-
-            var mockBlobItemResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockBlobItemResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var streamResult = (IEnumerable<Stream>)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsAssignableFrom<IEnumerable<Stream>>(streamResult);
-            Assert.Equal(expectedStream.ToString(), streamResult.First().ToString());
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_StreamCollection_Array_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(Stream[]), grpcModelBindingData);
-
-            var expectedStream = new MemoryStream(Encoding.UTF8.GetBytes("MyBlobStream"));
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadStreamingResult(expectedStream);
-
-            var mockResponse = new Mock<Response<BlobDownloadStreamingResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadStreamingAsync(It.IsAny<HttpRange>(), It.IsAny<BlobRequestConditions>(), It.IsAny<bool>(), null, default))
-                .ReturnsAsync(mockResponse.Object);
-
-            var mockBlobItemResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockBlobItemResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var streamResult = (Stream[])conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.Equal(expectedStream.ToString(), streamResult.First().ToString());
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_ByteArray_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(byte[]), grpcModelBindingData);
-
-            var expectedByteArray = Encoding.UTF8.GetBytes("MyBlobByteArray");
-            var testStream = new MemoryStream(expectedByteArray)
-            {
-                Position = 0
-            };
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadToAsync(It.IsAny<Stream>(), default))
-                .Callback<Stream, CancellationToken>(async (s, _) => await testStream.CopyToAsync(s))
-                .ReturnsAsync(new Mock<Response>().Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var byteResult = (byte[])conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsType<byte[]>(byteResult);
-            Assert.Equal(new byte[0], byteResult);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_ByteArrayCollection_List_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(IEnumerable<byte[]>), grpcModelBindingData);
-
-            var expectedByteArray = Encoding.UTF8.GetBytes("MyBlobByteArray");
-            var testStream = new MemoryStream(expectedByteArray);
-            testStream.Position = 0;
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadToAsync(It.IsAny<Stream>(), default))
-                .Callback<Stream, CancellationToken>(async (s, _) => await testStream.CopyToAsync(s))
-                .ReturnsAsync(new Mock<Response>().Object);
-
-            var mockResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var byteResult = (IEnumerable<byte[]>)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsAssignableFrom<IEnumerable<byte[]>>(byteResult);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_ByteArrayCollection_Array_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(byte[][]), grpcModelBindingData);
-
-            var expectedByteArray = Encoding.UTF8.GetBytes("MyBlobByteArray");
-            var testStream = new MemoryStream(expectedByteArray);
-            testStream.Position = 0;
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadToAsync(It.IsAny<Stream>(), default))
-                .Callback<Stream, CancellationToken>(async (s, _) => await testStream.CopyToAsync(s))
-                .ReturnsAsync(new Mock<Response>().Object);
-
-            var mockResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var byteResult = (byte[][])conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsType<byte[][]>(byteResult);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_POCO_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(Book), grpcModelBindingData);
-
-            var expectedBook = new Book() { Name = "MyBook" };
-
-            var testStream = new MemoryStream(Encoding.UTF8.GetBytes("{\"Name\":\"MyBook\"}"));
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadStreamingResult(testStream);
-            var mockResponse = new Mock<Response<BlobDownloadStreamingResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadStreamingAsync(It.IsAny<HttpRange>(), It.IsAny<BlobRequestConditions>(), It.IsAny<bool>(), null, default))
-                .ReturnsAsync(mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var pocoResult = (Book)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.Equal(expectedBook.GetType(), pocoResult.GetType());
-            Assert.Equal(expectedBook.Name, pocoResult.Name);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_POCOCollection_List_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(IEnumerable<Book>), grpcModelBindingData);
-
-            var expectedBookList = new List<Book>() { new Book() { Name = "MyBook" } };
-            var testStream = new MemoryStream(Encoding.UTF8.GetBytes("{\"Name\":\"MyBook\"}"));
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadStreamingResult(testStream);
-
-            var mockResponse = new Mock<Response<BlobDownloadStreamingResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadStreamingAsync(It.IsAny<HttpRange>(), It.IsAny<BlobRequestConditions>(), It.IsAny<bool>(), null, default))
-                .ReturnsAsync(mockResponse.Object);
-
-            var mockBlobItemResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockBlobItemResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var pocoResult = (IEnumerable<Book>)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsAssignableFrom<IEnumerable<Book>>(pocoResult);
-            Assert.Equal(expectedBookList[0].Name, pocoResult.First().Name);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ValidModelBindingData_POCOCollection_Array_ReturnsSuccess()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(Book[]), grpcModelBindingData);
-
-            var expectedBookList = new Book[] { new Book() { Name = "MyBook" } };
-            var testStream = new MemoryStream(Encoding.UTF8.GetBytes("{\"Name\":\"MyBook\"}"));
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadStreamingResult(testStream);
-
-            var mockResponse = new Mock<Response<BlobDownloadStreamingResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadStreamingAsync(It.IsAny<HttpRange>(), It.IsAny<BlobRequestConditions>(), It.IsAny<bool>(), null, default))
-                .ReturnsAsync(mockResponse.Object);
-
-            var mockBlobItemResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, mockBlobItemResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var pocoResult = (Book[])conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Succeeded, conversionResult.Status);
-            Assert.IsType<Book[]>(pocoResult);
-            Assert.Equal(expectedBookList[0].Name, pocoResult.First().Name);
-        }
-
-        // Unhappy cases
-
-        [Fact]
-        public async Task ConvertAsync_ContentSource_AsObject_ReturnsUnhandled()
-        {
-            // Arrange
-            var context = new TestConverterContext(typeof(BlobClient), new object());
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Unhandled, conversionResult.Status);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ModelBindingData_Null_ReturnsUnhandled()
-        {
-            // Arrange
-            var context = new TestConverterContext(typeof(BlobClient), null);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Unhandled, conversionResult.Status);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_BlobClientIsNull_ReturnsUnhandled()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns((BlobClient)null);
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.Contains("Value cannot be null", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ThrowsException_ReturnsFailure()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            _mockBlobServiceClient
-                .Setup(m => m.GetBlobContainerClient(It.IsAny<string>()))
-                .Throws(new Exception());
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ModelBindingDataSource_NotBlobExtension_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "anotherExtensions");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ModelBindingDataContentType_Unsupported_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs", contentType: "binary");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.Equal("Unexpected content-type. Only 'application/json' is supported.", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ModelBindingData_InvalidContent_Throws_ReturnsFailed()
-        {
-            // Arrange
-            string badJsonData = $@"{{
-                                ""Connection"" : ""Connection"",
-                                ""ContainerName"" ""ContainerName"",
-                                ""BlobName"" : ""BlobName"",
-                            }}";
-
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(new BinaryData(badJsonData), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Contains("Binding parameters to complex objects uses JSON serialization", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_MissingConnectionParameter_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(null), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var clientResult = (BlobClient)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.Equal("Value cannot be null. (Parameter 'Connection')", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_MissingContainerNameParameter_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(container: null), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var clientResult = (BlobClient)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.Equal("Value cannot be null. (Parameter 'ContainerName')", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_BlobClient_MissingBlobNameParameter_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: null), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobClient), grpcModelBindingData);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var clientResult = (BlobClient)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.Equal("Value cannot be null. (Parameter 'blobName')", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_POCO_InvalidJson_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(Book), grpcModelBindingData);
-
-            var expectedBook = new Book() { Name = "MyBook" };
-
-            var testStream = new MemoryStream(Encoding.UTF8.GetBytes("{\"Name:\"MyBook\"}"));
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadStreamingResult(testStream);
-            var mockResponse = new Mock<Response<BlobDownloadStreamingResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadStreamingAsync(It.IsAny<HttpRange>(), It.IsAny<BlobRequestConditions>(), It.IsAny<bool>(), null, default))
-                .ReturnsAsync(mockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-            var pocoResult = (Book)conversionResult.Value;
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Contains("Binding parameters to complex objects uses JSON serialization", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_IncorrectJsonContent_POCOCollection_Array_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(Book[]), grpcModelBindingData);
-
-            var expectedBookList = new Book[] { new Book() { Name = "MyBook" } };
-            var testStream = new MemoryStream(Encoding.UTF8.GetBytes("i should fail"));
-            var blobDownloadResult = BlobsModelFactory.BlobDownloadStreamingResult(testStream);
-
-            var mockResponse = new Mock<Response<BlobDownloadStreamingResult>>();
-            mockResponse.SetupGet(r => r.Value).Returns(blobDownloadResult);
-
-            var mockBlobClient = new Mock<BlobClient>();
-            mockBlobClient
-                .Setup(m => m.DownloadStreamingAsync(It.IsAny<HttpRange>(), It.IsAny<BlobRequestConditions>(), It.IsAny<bool>(), null, default))
-                .ReturnsAsync(mockResponse.Object);
-
-            var blobItemMockResponse = new Mock<Response>();
-            var expectedOutput = Page<BlobItem>.FromValues(new List<BlobItem>{ BlobsModelFactory.BlobItem("MyBlob") }, continuationToken: null, blobItemMockResponse.Object);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.GetBlobClient(It.IsAny<string>())).Returns(mockBlobClient.Object);
-            mockContainer.Setup(m => m.GetBlobsAsync(It.IsAny<BlobTraits>(), It.IsAny<BlobStates>(), It.IsAny<string>(), default))
-                            .Returns(AsyncPageable<BlobItem>.FromPages(new List<Page<BlobItem>> { expectedOutput }));
-
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Contains("Binding parameters to complex objects uses JSON serialization", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_BlobClient_CollectionBindingWithBlobName_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(IEnumerable<BlobClient>), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Equal("Collections are not supported when binding to a specific blob file.", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_StringCollection_CollectionBindingWithBlobName_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(string[]), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Equal("Collections are not supported when binding to a specific blob file.", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_Stream_CollectionBindingWithBlobName_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(IEnumerable<Stream>), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Equal("Collections are not supported when binding to a specific blob file.", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_ByteArray_CollectionBindingWithBlobName_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(byte[][]), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Equal("Collections are not supported when binding to a specific blob file.", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_POCO_CollectionBindingWithBlobName_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(Book[]), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Equal("Collections are not supported when binding to a specific blob file.", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_BlobContainerClient_BindingWithBlobName_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(blobName: "MyBlob.txt"), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(BlobContainerClient), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.Name).Returns("MyContainer");
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Contains("Binding to BlobContainerClient with a blob path is not supported", conversionResult.Error.Message);
-        }
-
-        [Fact]
-        public async Task ConvertAsync_BlobContainerClient_CollectionBinding_ReturnsFailed()
-        {
-            // Arrange
-            var grpcModelBindingData = Helper.GetTestGrpcModelBindingData(GetTestBinaryData(), "AzureStorageBlobs");
-            var context = new TestConverterContext(typeof(IEnumerable<BlobContainerClient>), grpcModelBindingData);
-
-            var mockContainer = new Mock<BlobContainerClient>();
-            mockContainer.Setup(m => m.Name).Returns("MyContainer");
-
-            _mockBlobServiceClient.Setup(m => m.GetBlobContainerClient(It.IsAny<string>())).Returns(mockContainer.Object);
-
-            // Act
-            var conversionResult = await _blobStorageConverter.ConvertAsync(context);
-
-            // Assert
-            Assert.Equal(ConversionStatus.Failed, conversionResult.Status);
-            Assert.IsType<InvalidOperationException>(conversionResult.Error);
-            Assert.Equal("Collections of BlobContainerClient are not supported.", conversionResult.Error.Message);
-        }
-
-        private BinaryData GetTestBinaryData(string connection = "Connection", string container = "Container", string blobName = "")
-        {
-            string jsonData = $@"{{
-                                ""Connection"" : ""{connection}"",
-                                ""ContainerName"" : ""{container}"",
-                                ""BlobName"" : ""{blobName}""
-                            }}";
-
-            return new BinaryData(jsonData);
-        }
-
-        public class Book
-        {
-            public string Name { get; set; }
-        }
-
-        public interface TMock
-        {
-        }
-    }
-}
